feat: rename settings and users page
